refactor: compare batched and per-note MIDI animation
